requirements.txt
Django==4.2.4
djangorestframework==3.14.0
joblib==1.3.0
numpy==1.23.4
gunicorn==20.1.0
